+++ xknx/exceptions/exception.py
@@ -21,13 +21,17 @@ class XKNXException(Exception):
 
 
 class CommunicationError(XKNXException):
-    """Unable to communicate with KNX Bus."""
+    """Unable to communicate with KNX bus."""
 
     def __init__(self, message: str, should_log: bool = True) -> None:
         """Instantiate exception."""
         super().__init__(message)
 
         self.should_log = should_log
+
+
+class ConfirmationError(CommunicationError):
+    """No confirmation received from KNX server for sent Telegram."""
 
 
 class CouldNotParseTelegram(XKNXException):
@@ -161,3 +165,15 @@ class InterfaceWithUserIdNotFound(SecureException):
 
 class InvalidSecureConfiguration(SecureException):
     """Exception class used when the secure configuration is invalid."""
+
+
+class ManagementConnectionError(XKNXException):
+    """Exception class used when a management connection fails."""
+
+
+class ManagementConnectionRefused(ManagementConnectionError):
+    """Exception class used when a management connection request is refused."""
+
+
+class ManagementConnectionTimeout(ManagementConnectionError):
+    """Exception class used when a management connection timed out."""