+++ xknx/management/__init__.py
@@ -1,0 +1,3 @@
+"""Package for management procedures as described in KNX-Standard 3.5.2."""
+# flake8: noqa
+from .management import Management